moved supportedCurrencies from advanced-logic to currency package
