+++ packages/advanced-logic/src/extensions/payment-network/any-to-erc20-proxy.ts
@@ -1,6 +1,6 @@
 import { ExtensionTypes, RequestLogicTypes } from '@requestnetwork/types';
 import Erc20FeeProxyPaymentNetwork from './erc20/fee-proxy-contract';
-import { supportedCurrencies } from './conversion-supported-currencies';
+import { supportedCurrencies } from '@requestnetwork/currency';
 
 const CURRENT_VERSION = '0.1.0';
